+++ fiddles/javac/fiddle-0001-LogForging/logforging/src/Fiddle.java
@@ -1,6 +1,0 @@
-
-class Fiddle {
-    public static void main(String[] args) {
-        System.out.println("logforging compiled!");
-    }
-}